CONTRIBUTING: clarify that tests are required
See https://github.com/burrowers/garble/issues/121#issuecomment-695935859.

In some rare cases, it's nearly impossible to write a test for a change,
but they are truly so rare that we shouldn't give any ideas here.

By default, all contributors should try to write a test for every
change that changes what the code is meant to do.